move SplitFullName to orders
[#123985341]
