+++ utils/generic.go
@@ -337,24 +337,3 @@ func StrToCamelCase(str string) string {
 
 	return str
 }
-
-// SplitFullName will take a fullname as a string and split it into first name and last names
-func SplitFullName(name string) (string, string) {
-
-	var firstName, lastName string
-
-	fullName := strings.SplitN(name, " ", 2)
-
-	if len(fullName) == 2 {
-		firstName = fullName[0]
-		lastName = fullName[1]
-	} else if len(fullName) == 1 {
-		firstName = fullName[0]
-		lastName = ""
-	} else {
-		firstName = ""
-		lastName = ""
-	}
-
-	return firstName, lastName
-}
